removed some old text and simplified
removed text:
<p>
Our three-man team has developed a sonification system for controlling various aspects of an electronic music performance. We use continuous inputs such as respiration to control synthesis and effect parameters. Events such as heartbeats provide a sync source for tempo and trigger scenes, samples and MIDI notes. Accelerometer-based posture measurement is typically used so that lateral movements alter the melodic structure of a piece, leaning forward is interpreted as an aggressive posture (reflected in the song’s dynamics), and leaning backwards eases the tension in the song. Bio-controlled lighting and projected visualisations are generated to augment the musical piece.
</p>

<p>
Performing electronic music live is problematic, as it is either pre-programmed or algorithmic in nature. A deeper level of interfacing with machines is required for sufficiently expressive performances. 
</p>